requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
alfworld>=0.2.2
ortools>=9.6.2534
pyperplan>=2.1.0
numpy<2.0.0 # Strict pinning for alfworld compat
pillow>=9.0.0
Pillow>=9.0.0
pandas>=2.0.0
